+++ Notes/reference/releaseNotes.md
@@ -3,6 +3,12 @@
 Some basic information on our tagged releases.
 Note - `git log tag1...tag2` shows the commit log between versions.
 
+## 1.4.5
+
+* bump versions - pickup some fixes
+* add config-.json files
+* analytics load site-id from appcontext config
+
 ## 1.4.4
 
 * add answers to common interview questions to resume